pyproject.toml
[tool.poetry]
name = "llama_2_document_answering"
version = "0.1.0"
description = ""
authors = ["Aizenit Technologies"]
license = ""
readme = "README.md"

[tool.poetry.scripts]
start = "main:app"

[tool.poetry.dependencies]
python = "^3.9"
ctransformers = "^0.2.5"
faiss-cpu = "1.7.4"
fastapi = "^0.96.0"
ipykernel = "^6.23.1"
langchain = "0.0.225"
pypdf = "3.8.1"
python-box = "7.0.1"
python-dotenv = "1.0.0"
sentence-transformers = "2.2.2"
uvicorn = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
